+++ Packages/com.tk.haptics/Runtime/NullHapticBackend.cs
@@ -5,7 +5,6 @@ namespace TK.Haptics
     {
         public bool IsSupported => false;
         public bool SystemTouchVibrationDisabled => false;
-        public bool BypassSystemVibrationSetting { get; set; }   // stored, inert
         public void Impact(HapticImpact strength) { }
         public void Selection() { }
         public void Notification(HapticNotification type) { }